most packets are done!

+++ GamerellaJam/src/com/envy3d/shared/Entity.java
@@ -1,12 +1,12 @@
 package com.envy3d.shared;
 
 public class Entity {
-	public short id;
+	public int id;
 	private boolean gilded;
 	public int posX, posY;
 	public IntVec2 dest;
 	
-	public Entity(short id, boolean gilded, int posX, int posY) {
+	public Entity(int id, boolean gilded, int posX, int posY) {
 		this.id = id;
 		this.gilded = gilded;
 		this.posX = posX;
